Renamed `enableSourceControl` to `sourceControlIsEnabled`. Added documentation and TODO.
