+++ CodeEdit/Features/CodeEditUI/Views/ToolbarBranchPicker.swift
@@ -84,7 +84,7 @@ struct ToolbarBranchPicker: View {
             self.currentBranch = branch
         }
         .task {
-            if Settings.shared.preferences.sourceControl.general.enableSourceControl {
+            if Settings.shared.preferences.sourceControl.general.sourceControlIsEnabled {
                 await self.sourceControlManager?.refreshCurrentBranch()
                 await self.sourceControlManager?.refreshBranches()
             }
