+++ CodeEdit/Features/Settings/Pages/SourceControlSettings/SourceControlGeneralView.swift
@@ -19,7 +19,7 @@ struct SourceControlGeneralView: View {
     var body: some View {
         SettingsForm {
             Section("Source Control") {
-                enableSourceControl
+                sourceControlIsEnabled
                 refreshLocalStatusAuto
                 fetchRefreshStatusAuto
                 addRemoveFilesAuto
@@ -47,10 +47,10 @@ struct SourceControlGeneralView: View {
 }
 
 private extension SourceControlGeneralView {
-    private var enableSourceControl: some View {
+    private var sourceControlIsEnabled: some View {
         Toggle(
             "Enable source control",
-            isOn: $settings.enableSourceControl
+            isOn: $settings.sourceControlIsEnabled
         )
     }
 
@@ -59,7 +59,7 @@ private extension SourceControlGeneralView {
             "Refresh local status automatically",
             isOn: $settings.refreshStatusLocally
         )
-        .disabled(!settings.enableSourceControl)
+        .disabled(!settings.sourceControlIsEnabled)
     }
 
     private var fetchRefreshStatusAuto: some View {
@@ -67,7 +67,7 @@ private extension SourceControlGeneralView {
             "Fetch and refresh server status automatically",
             isOn: $settings.fetchRefreshServerStatus
         )
-        .disabled(!settings.enableSourceControl)
+        .disabled(!settings.sourceControlIsEnabled)
     }
 
     private var addRemoveFilesAuto: some View {
@@ -75,7 +75,7 @@ private extension SourceControlGeneralView {
             "Add and remove files automatically",
             isOn: $settings.addRemoveAutomatically
         )
-        .disabled(!settings.enableSourceControl)
+        .disabled(!settings.sourceControlIsEnabled)
     }
 
     private var selectFilesToCommitAuto: some View {
@@ -83,7 +83,7 @@ private extension SourceControlGeneralView {
             "Select files to commit automatically",
             isOn: $settings.selectFilesToCommit
         )
-        .disabled(!settings.enableSourceControl)
+        .disabled(!settings.sourceControlIsEnabled)
     }
 
     private var showSourceControlChanges: some View {
@@ -91,7 +91,7 @@ private extension SourceControlGeneralView {
             "Show source control changes",
             isOn: $settings.showSourceControlChanges
         )
-        .disabled(!settings.enableSourceControl)
+        .disabled(!settings.sourceControlIsEnabled)
     }
 
     private var includeUpstreamChanges: some View {
@@ -99,7 +99,7 @@ private extension SourceControlGeneralView {
             "Include upstream changes",
             isOn: $settings.includeUpstreamChanges
         )
-        .disabled(!settings.enableSourceControl || !settings.showSourceControlChanges)
+        .disabled(!settings.sourceControlIsEnabled || !settings.showSourceControlChanges)
     }
 
     private var comparisonView: some View {
